This is the current state of play
What I've got so far is a webpage (questionbank1.html and questionbank1.js) that will load a question and check an answer. The next step would be to use the (partial) question bank (questionbank.csv) onto a database for it to call from, rather than loading it all in the js file. After that I would be looking to have it refer to a students' account and where they are in the course to automatically give them a question using a bit of code similar to the firstpython.py file. Eventually I would like this as an app that student's can have on their phones.